server/requirements.txt
# server/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
Pillow
python-multipart